+++ src/Avalonia.Controls.Maui/Platform/MauiAvaloniaWindow.cs
@@ -7,7 +7,6 @@ using Microsoft.Maui;
 using Microsoft.Maui.Platform;
 using System;
 using System.Collections.Generic;
-using System.Threading.Tasks;
 
 namespace Avalonia.Controls.Maui.Platform;
 
@@ -21,7 +20,7 @@ namespace Avalonia.Controls.Maui.Platform;
 /// </remarks>
 public class MauiAvaloniaWindow : Window, IDisposable
 {
-    private readonly Stack<(Control Content, Panel Scrim)> _modalStack = new();
+    private readonly ModalOverlayHelper _modalHelper;
     private readonly DockPanel _rootPanel;
     private readonly Grid _rootGrid;
     private Control? _mainContent;
@@ -50,6 +49,11 @@ public class MauiAvaloniaWindow : Window, IDisposable
         // the entire visual tree and trigger spurious MAUI Unloaded/Loaded events
         // that reset Shell state (e.g. hamburger button visibility).
         _rootGrid = new Grid { Tag = "OverlayWrapper" };
+        _modalHelper = new ModalOverlayHelper(_rootGrid, () =>
+        {
+            var h = Bounds.Height;
+            return h > 0 ? h : ClientSize.Height;
+        });
         _rootGrid.Children.Add(_rootPanel);
 
         Content = _rootGrid;
@@ -199,107 +203,17 @@ public class MauiAvaloniaWindow : Window, IDisposable
     /// Presents a modal page as a fullscreen overlay within the same window.
     /// Covers all shell chrome (flyout, navbar, tabs) for a true modal experience.
     /// </summary>
-    public async void PresentModal(Control modalContent, bool animated = true)
+    public void PresentModal(Control modalContent, bool animated = true)
     {
-        int zBase = 100 + _modalStack.Count * 2;
-
-        var scrim = new Panel
-        {
-            Background = new SolidColorBrush(Colors.Black, 0.5),
-            ZIndex = zBase,
-            IsHitTestVisible = true,
-            Opacity = animated ? 0 : 1
-        };
-
-        // Wrap modal content in a container with an opaque background so it
-        // fully covers the shell chrome beneath (flyout, navbar, tabs).
-        // The page's own BackgroundColor (if set) will paint over this default.
-        var theme = Application.Current?.ActualThemeVariant;
-        var defaultBg = theme == Avalonia.Styling.ThemeVariant.Dark
-            ? new SolidColorBrush(Color.Parse("#1F1F1F"))
-            : (IBrush)Brushes.White;
-
-        var wrapper = new Border
-        {
-            Child = modalContent,
-            Background = defaultBg,
-            ZIndex = zBase + 1,
-            HorizontalAlignment = HorizontalAlignment.Stretch,
-            VerticalAlignment = VerticalAlignment.Stretch
-        };
-
-        _modalStack.Push((wrapper, scrim));
-        _rootGrid.Children.Add(scrim);
-        _rootGrid.Children.Add(wrapper);
-
-        if (animated)
-        {
-            var windowHeight = Bounds.Height > 0 ? Bounds.Height : ClientSize.Height;
-            if (windowHeight <= 0) windowHeight = 800;
-
-            var transform = new TranslateTransform(0, windowHeight);
-            wrapper.RenderTransform = transform;
-
-            var duration = TimeSpan.FromMilliseconds(400);
-            var startTime = DateTime.Now;
-
-            while (DateTime.Now - startTime < duration)
-            {
-                var progress = (DateTime.Now - startTime).TotalMilliseconds / duration.TotalMilliseconds;
-                progress = Math.Min(1.0, progress);
-                progress = 1 - Math.Pow(1 - progress, 3); // ease out cubic
-
-                transform.Y = windowHeight * (1 - progress);
-                scrim.Opacity = progress;
-                await Task.Delay(16);
-            }
-
-            transform.Y = 0;
-            scrim.Opacity = 1;
-            wrapper.RenderTransform = null;
-        }
+        _modalHelper.Present(modalContent, animated);
     }
 
     /// <summary>
     /// Dismisses the top-most modal overlay.
     /// </summary>
-    public async void DismissModal(bool animated = true)
+    public void DismissModal(bool animated = true)
     {
-        if (_modalStack.Count == 0)
-            return;
-
-        var (modalContent, scrim) = _modalStack.Pop();
-
-        try
-        {
-            if (animated)
-            {
-                var windowHeight = Bounds.Height > 0 ? Bounds.Height : ClientSize.Height;
-                if (windowHeight <= 0) windowHeight = 800;
-
-                var transform = new TranslateTransform(0, 0);
-                modalContent.RenderTransform = transform;
-
-                var duration = TimeSpan.FromMilliseconds(400);
-                var startTime = DateTime.Now;
-
-                while (DateTime.Now - startTime < duration)
-                {
-                    var progress = (DateTime.Now - startTime).TotalMilliseconds / duration.TotalMilliseconds;
-                    progress = Math.Min(1.0, progress);
-                    progress = 1 - Math.Pow(1 - progress, 3); // ease out cubic
-
-                    transform.Y = windowHeight * progress;
-                    scrim.Opacity = 1 - progress;
-                    await Task.Delay(16);
-                }
-            }
-        }
-        finally
-        {
-            _rootGrid.Children.Remove(modalContent);
-            _rootGrid.Children.Remove(scrim);
-        }
+        _modalHelper.Dismiss(animated);
     }
 
     /// <summary>
@@ -312,16 +226,11 @@ public class MauiAvaloniaWindow : Window, IDisposable
     /// </remarks>
     public void ClearAllModals()
     {
-        while (_modalStack.Count > 0)
-        {
-            var (modalContent, scrim) = _modalStack.Pop();
-            _rootGrid.Children.Remove(modalContent);
-            _rootGrid.Children.Remove(scrim);
-        }
+        _modalHelper.ClearAll();
     }
 
     /// <summary>
     /// Gets the current number of presented modals.
     /// </summary>
-    public int ModalCount => _modalStack.Count;
+    public int ModalCount => _modalHelper.Count;
 }
